trying out now actions speed dial
